% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerMCMC-package.R
\docType{package}
\name{ConquerMCMC-package}
\alias{ConquerMCMC}
\alias{ConquerMCMC-package}
\title{ConquerMCMC: Divide and Conquer Algorithms for Markov Chain Monte Carlo}
\description{
A series of implementations of divide-and-conquer algorithms for Markov-Chain Monte Carlo. These functions provide scalable Metropolis-Hastings samplers to deal with large and disjointed data sets. Three methods, excluding the standard MCMC scheme, are available: the consensus MCMC, the m-posterior, and the scaled subprior. See function descriptions for proper attributions.
}
\author{
\strong{Maintainer}: Jacob Raymond \email{j4raymond@uwaterloo.ca}

}
\keyword{internal}
