% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{mcmc}
\alias{mcmc}
\title{Metropolis-Hastings Sampler}
\usage{
mcmc(
  para,
  startval,
  niter,
  X,
  prior,
  likelihood,
  accept = F,
  propvar = NULL,
  burn.rate = 0.1
)
}
\arguments{
\item{para}{Parameters to be estimated.}

\item{startval}{Initial value of the chain.}

\item{niter}{Number of iterations (including burned iterations).}

\item{X}{Matrix of observations from the underlying model.}

\item{prior}{Prior distribution for the parameters.}

\item{likelihood}{Likelihood function.}

\item{accept}{Logical. If true, displays the acceptance rate of the chain.}

\item{propvar}{The diagonal of the matrix variance for the proposal distribution. If no value is specified, the identify matrix is used.}

\item{burn.rate}{The percentage of iterations to be burned.}
}
\value{
A list with the following items:
\describe{
\item{\code{Chains}}{A dataframe containing the chain iterations that were not burned.}
\item{\code{Estimates}}{The parameter estimates.}
\item{\code{Log-Likelihood}}{A dataframe containing the value of the likelihood function and prior evaluated at each point.}
}
}
\description{
A basic Metropolis-Hastings Ssmpler for parameter estimation. Utilizes a Metropolis random walk transition density.
}
\examples{
#Parameter Estimation for Data from a Normal Distribution

#Prior
prior<-function(param){
  ifelse(all(param>0), 1, 0)
}

#Likelihood function
normal.likelihood<-function(X, param){
  mu=param[1]
  sigma=param[2]
  sum(dnorm(x=X, mean=mu, sd=sigma, log=TRUE))
}

#Simulate data
X<-rnorm(100, 2, 1.2)

#Parameters
param<-c("mu", "sigma")
niter<-10000
startval<-c(1, 1)

df<-mcmc(param, startval, niter, X, prior, normal.likelihood, accept=TRUE, propvar=0.25)
}
