% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workers.R
\name{workers}
\alias{workers}
\title{Segment a data set into subsets}
\usage{
workers(X, n, random = T)
}
\arguments{
\item{X}{A matrix to be subsetted.}

\item{n}{Number of subsets to be created.}

\item{random}{If true, the rows of X are shuffled prior to the split.}
}
\value{
A list containing the n subsets.
}
\description{
This algorithm segments a data set into a prespecified number of subsets.
}
\examples{
#Split a matrix with 100 rows and two columns into 5 subsets

#Simulate data
df<-matrix(data = c(rnorm(100), runif(100)), ncol=2)

#Return list of subsets
workers(df, n=5)

}
